Moved state from App Component to Redux Store & connected App component to Redux store.
