+++ src/App.js
@@ -5,31 +5,11 @@ import AddedFeatures from "./components/AddedFeatures";
 import AdditionalFeatures from "./components/AdditionalFeatures";
 import Total from "./components/Total";
 
-import { createStore } from "redux";
-import { Provider } from "react-redux";
-import { appReducer } from "./reducers/appReducer";
+import { connect } from "react-redux";
 
-const store = createStore(appReducer);
 // console.log("store", store);
 
-const App = () => {
-	const state = {
-		additionalPrice: 0,
-		car: {
-			price: 26395,
-			name: "2019 Ford Mustang",
-			image:
-				"https://cdn.motor1.com/images/mgl/0AN2V/s1/2019-ford-mustang-bullitt.jpg",
-			features: []
-		},
-		additionalFeatures: [
-			{ id: 1, name: "V-6 engine", price: 1500 },
-			{ id: 2, name: "Racing detail package", price: 1500 },
-			{ id: 3, name: "Premium sound system", price: 500 },
-			{ id: 4, name: "Rear spoiler", price: 250 }
-		]
-	};
-
+const App = props => {
 	const removeFeature = item => {
 		// dispatch an action here to remove an item
 	};
@@ -39,19 +19,25 @@ const App = () => {
 	};
 
 	return (
-		<Provider store={store}>
-			<div className='boxes'>
-				<div className='box'>
-					<Header car={state.car} />
-					<AddedFeatures car={state.car} />
-				</div>
-				<div className='box'>
-					<AdditionalFeatures additionalFeatures={state.additionalFeatures} />
-					<Total car={state.car} additionalPrice={state.additionalPrice} />
-				</div>
+		<div className='boxes'>
+			<div className='box'>
+				<Header car={props.car} />
+				<AddedFeatures car={props.car} />
 			</div>
-		</Provider>
+			<div className='box'>
+				<AdditionalFeatures additionalFeatures={props.additionalFeatures} />
+				<Total car={props.car} additionalPrice={props.additionalPrice} />
+			</div>
+		</div>
 	);
 };
 
-export default App;
+const mapStateToProps = state => {
+	return {
+		additionalPrice: state.additionalPrice,
+		car: state.car,
+		additionalFeatures: state.additionalFeatures
+	};
+};
+
+export default connect(mapStateToProps, {})(App);
